Turn song progress bar to a seek bar
Seek bar replaces the song progress bar.
Adds a listener to the new seek bar for the user to change the song's current time.
Adds checks to stop song stuttering from the timer updating the seek bar itself.
Adds a basic style to the seek bar to differentiate it from the volume slider.